+++ tests/TestApp/application/models/Comment.php
@@ -2,4 +2,5 @@
 
 class Comment extends Rougin\Wildfire\CodeigniterModel
 {
+	protected $hidden = array('id');
 }